Rename `conway` to `game_of_life` for clarity
The file that holds the game of life logic has been named similarly for
clarity
