+++ game_of_life.py
